--- auth-service/src/main/java/mx/com/akkihazlo/microservice/authservice/AuthServiceApplication.java
@@ -1,14 +1,0 @@
-package mx.com.akkihazlo.microservice.authservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class AuthServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AuthServiceApplication.class, args);
-	}
-}
