--- config-service/src/main/java/mx/com/akkihazlo/microservice/configservice/ConfigServiceApplication.java
@@ -1,16 +1,0 @@
-package mx.com.akkihazlo.microservice.configservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.config.server.EnableConfigServer;
-
-@SpringBootApplication
-@EnableConfigServer
-@EnableDiscoveryClient
-public class ConfigServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ConfigServiceApplication.class, args);
-	}
-}
