--- config-service/src/test/java/mx/com/akkihazlo/microservice/configservice/ConfigServiceApplicationTests.java
@@ -1,16 +1,0 @@
-package mx.com.akkihazlo.microservice.configservice;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class ConfigServiceApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
