--- eureka-service/src/main/java/mx/com/akkihazlo/microservice/eurekaservice/EurekaServiceApplication.java
@@ -1,14 +1,0 @@
-package mx.com.akkihazlo.microservice.eurekaservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@SpringBootApplication
-@EnableEurekaServer
-public class EurekaServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EurekaServiceApplication.class, args);
-	}
-}
