--- gateway-service/src/main/java/mx/com/akkihazlo/gateway/GatewayServiceApplication.java
@@ -1,16 +1,0 @@
-package mx.com.akkihazlo.gateway;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-@EnableZuulProxy
-public class GatewayServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(GatewayServiceApplication.class, args);
-	}
-}
