+++ infrastructure-services/api-gateway/src/main/resources/bootstrap.yml
