+++ infrastructure-services/authentication-service/src/main/java/mx/com/akkihazlo/auth/AuthServiceApplication.java
@@ -1,0 +1,14 @@
+package mx.com.akkihazlo.auth;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class AuthServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AuthServiceApplication.class, args);
+	}
+}
