+++ infrastructure-services/authentication-service/src/main/resources/bootstrap.yml
