+++ infrastructure-services/configuration-service/src/main/java/mx/com/akkihazlo/configservice/ConfigServiceApplication.java
@@ -1,0 +1,17 @@
+package mx.com.akkihazlo.configservice;
+
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.config.server.EnableConfigServer;
+
+@SpringBootApplication
+@EnableConfigServer
+@EnableDiscoveryClient
+public class ConfigServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ConfigServiceApplication.class, args);
+	}
+}
