+++ infrastructure-services/discovery-service/src/main/java/mx/com/akkihazlo/discovery/DiscoveryServiceApplication.java
@@ -1,0 +1,14 @@
+package mx.com.akkihazlo.discovery;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
+
+@SpringBootApplication
+@EnableEurekaServer
+public class DiscoveryServiceApplication {
+	
+	public static void main(String[] args) {
+		SpringApplication.run(DiscoveryServiceApplication.class, args);
+	}
+}
