+++ infrastructure-services/discovery-service/src/main/resources/bootstrap.yml
